Major re-organization of service clients, protocol defaults
