+++ colossus-testkit/src/main/scala/colossus-testkit/HttpServiceSpec.scala
@@ -5,6 +5,7 @@ import scala.concurrent.Await
 import akka.util.ByteString
 
 import protocols.http._
+import Http.defaults._
 
 abstract class HttpServiceSpec extends ServiceSpec[Http] {
