+++ colossus-testkit/src/main/scala/colossus-testkit/ServiceSpec.scala
@@ -33,7 +33,7 @@ abstract class ServiceSpec[C <: CodecDSL](implicit provider: CodecProvider[C], c
     requestTimeout = timeout
   )
 
-  def client(timeout: FiniteDuration = requestTimeout) = AsyncServiceClient(clientConfig(timeout), clientProvider.clientCodec)
+  def client(timeout: FiniteDuration = requestTimeout) = AsyncServiceClient(clientConfig(timeout))//, clientProvider.clientCodec)
 
   def withClient(f: AsyncServiceClient[Request, Response] => Unit) {
     val c = client()
