+++ colossus-tests/src/it/scala/colossus/MemcacheITSpec.scala
@@ -3,7 +3,7 @@ package colossus
 import java.net.InetSocketAddress
 
 import akka.util.ByteString
-import colossus.protocols.memcache.MemcacheClient
+import colossus.protocols.memcache._
 import colossus.protocols.memcache.MemcacheReply._
 import colossus.protocols.memcache.{MemcacheCommand, MemcacheReply}
 import colossus.service.{AsyncServiceClient, ClientConfig}
@@ -14,6 +14,9 @@ import org.scalatest.time.{Millis, Seconds, Span}
 import scala.collection.mutable
 import scala.concurrent.{Await, Future}
 import scala.concurrent.duration._
+
+import Memcache.defaults._
+
 /*
 Please be aware when running this test, that if you run it on a machine with a memcached server
 running on 11211, it will begin to interact with it.  Be mindful if you are running this on a server
@@ -29,7 +32,7 @@ class MemcacheITSpec extends ColossusSpec with ScalaFutures{
 
   implicit val sys = IOSystem("test-system", 2)
 
-  val client = MemcacheClient.asyncClient(ClientConfig(new InetSocketAddress("localhost", 11211), 2.seconds, "memcache"))
+  val client = Memcache.futureClient(ClientConfig(new InetSocketAddress("localhost", 11211), 2.seconds, "memcache"))
 
   val usedKeys = scala.collection.mutable.HashSet[ByteString]()
